pyproject.toml
[project]
name = "extropic-hack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jax[cuda12]>=0.6.2",
    "thrml>=0.1.3",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
]
